+++ src/es/ubu/inf/tfg/asientosContables/CompraMercaderias.java
@@ -7,7 +7,7 @@ import es.ubu.inf.tfg.otrasCosas.Enunciado;
 
 public class CompraMercaderias extends Asiento{
 
-	public CompraMercaderias(Calendar f, int[] i) {
+	public CompraMercaderias(Calendar f, double[] i) {
 		fecha =f;
 		inputs=i;
 		
@@ -24,7 +24,7 @@ public class CompraMercaderias extends Asiento{
 				
 		//SE SALDAN LAS DEUDAS CON LOS PROVEEDORES "Y" DIAS DESPUES:
 		Calendar fechaDeudas = (Calendar)fecha.clone();
-		fechaDeudas.add(Calendar.DAY_OF_YEAR, +inputs[2]);
+		fechaDeudas.add(Calendar.DAY_OF_YEAR, (int) +inputs[2]);
 		
 		String enunciado2 = " Se salda la deuda con los vendedores de las mercaderías.\n"
 				+ "CUENTAS PGC: 400. Proveedores; 572. Bancos e instituciones de crédito c/c vista, euros. \n";
